# core/api/swagger/swagger.yaml
openapi: 3.0.0
info:
  description: |
    # Overview
    ## Key Concepts
    TODO

    ## Authentication
    TODO

    ## Representations
    TODO

    ## CORS
    TODO

    ## Versioning
    TODO
  title: Flagbase API
  version: 1.0.0
  contact:
    name: Flagbase
    url: 'https://flagbase.com'
    email: info@flagbase.com
  license:
    name: MPL 2.0
    url: 'https://flagbase.com/legal/license'
  x-logo:
    altText: Flagbase Core API
  termsOfService: 'https://flagbase.com/legal/privacy'
...
paths:
...
  '/projects/{wsKey}/{projKey}/environments/{envKey}/sdk-keys':
    parameters:
      - schema:
          type: string
        name: wsKey
        in: path
        required: true
        description: The workspace key
      - schema:
          type: string
        name: projKey
        in: path
        required: true
        description: The project key
      - schema:
          type: string
        name: envKey
        in: path
        required: true
        description: The environment key
    get:
      summary: List SDK keys
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      operationId: list-sdk-keys
      description: Retrieve a list of sdk keys for that particular environment
      tags:
        - sdk-keys
    post:
      summary: Create SDK key
      operationId: create-sdk-key
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      description: Create a sdk key for a particular project environment
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SDKKeyRequest'
        description: SDK Key request body
      tags:
        - sdk-keys
  '/projects/{wsKey}/{projKey}/environments/{envKey}/sdk-keys/{id}':
    parameters:
      - schema:
          type: string
        name: wsKey
        in: path
        required: true
        description: The workspace key
      - schema:
          type: string
        name: projKey
        in: path
        required: true
        description: The project key
      - schema:
          type: string
        name: envKey
        in: path
        required: true
        description: The environment key
      - schema:
          type: string
        name: id
        in: path
        required: true
        description: The SDK key id
    get:
      summary: List SDK keys
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      operationId: get-sdk-key
      description: Retrieve a list of sdk keys for that particular environment
      tags:
        - sdk-keys
    patch:
      summary: Update SDK Key
      operationId: update-sdk-key
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      description: Update SDK Key
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchDocument'
        description: Update sdk key
      tags:
        - sdk-keys
    delete:
      summary: Delete SDK Key
      operationId: delete-sdk-key
      responses:
        '204':
          description: No Content
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      description: Delete SDK key
      tags:
        - sdk-keys
servers:
  - url: 'https://api.flagbase.io'
    description: Flagbase Cloud
  - url: 'http://localhost:5051'
    description: Localhost
components:
...
  schemas:
    PatchDocument:
      description: A JSONPatch document as defined by RFC 6902
      title: PatchDocument
      type: array
      items:
        type: object
        properties:
          op:
            type: string
            description: The operation to be performed
            enum:
              - add
              - remove
              - replace
              - move
              - copy
              - test
          path:
            type: string
            description: A JSON-Pointer
          value:
            type: object
            description: The value to be used within the operations.
          from:
            type: string
            description: A string containing a JSON Pointer value.
...
    ResouceName:
      type: string
      title: Resouce Name
      description: A name used to label resources. Ideally a human-readable version of the resource key.
      maxLength: 30
      example: Example Name
    ResourceDescription:
      type: string
      title: Resource Description
      description: Optional resource description.
      example: This should be a description.
    ResourceTags:
      title: Resource Tags
      description: 'Optional tags used to identify this resource. '
      type: array
      x-examples: {}
      items:
        type: string
        example: example-tag
...
    SuccessResponse:
      title: SuccessResponse
      type: object
      properties:
        data:
          type: object
      description: Standard success response object
    ErrorResponse:
      title: ErrorResponse
      type: object
      description: Error response is an array of error objects
      x-examples: {}
      properties:
        errors:
          type: array
          items:
            $ref: '#/components/schemas/Error'
      required:
        - errors
...
    Error:
      title: Error
      type: object
      properties:
        code:
          type: string
        'message ':
          type: string
      description: 'Standard error response object, used inside an error response'
...
    SDKKeyRequest:
      title: SDKKeyRequest
      type: object
      description: SDK key request body
      properties:
        enabled:
          type: boolean
          description: Whether or not this SDK key is active
        expiresAt:
          type: number
          description: 'Time this key is no longer valid (unix timestamp) '
        name:
          $ref: '#/components/schemas/ResouceName'
        tags:
          $ref: '#/components/schemas/ResourceTags'
        description:
          $ref: '#/components/schemas/ResourceDescription'
      x-tags:
        - sdk-keys
...